refactor(app): extract diagnostic-jump orchestrators to diagnostic_jumps (phase 3e)
Move jump_to_next_error and jump_to_previous_error out of render.rs
into a new app/diagnostic_jumps.rs sibling. ~135 lines.

These walk the active buffer's diagnostic overlays in cursor order and
emit a MoveCursor event with a status message describing the diagnostic
at the new position.

render.rs goes from 3051 -> 2916 lines (-135). 2026 tests still pass.
